fixed method calculateAverage

+++ BuggyCodeExample.java
@@ -24,7 +24,7 @@ public class BuggyCodeExample {
         for (int i = 0; i < numbers.length; i++) {
             sum += numbers[i];
         }
-        return sum / numbers.length;
+        return (double) sum / numbers.length;
     }
 
     // 4. Main method to test the above methods
